// apps/api/src/thesis-applications/enums/thesis_status.enum.ts
export enum ThesisStatusEnum {
  APPLICANT = 'Applicant',
  POSTULADO = 'Postulado',
  ACEPTADO = 'Aceptado',
  INSCRITO = 'Inscrito',
  INFORME = 'Informe',
  FINALIZADO = 'Finalizado',
  SUBAREA_PENDING_ADVISOR = 'Por aprobar por el Asesor',
  SUBAREA_PENDING_COORDINATOR = 'Por aprobar por el Coordinador de Subárea',
  SUBAREA_APPROVED = 'Inscripción aprobada',
  REJECTED = 'Rechazado',
  THESIS1_PENDING_ADVISOR = 'Tesis 1: por aprobar por el Asesor',
  THESIS1_PENDING_COORDINATOR = 'Tesis 1: por aprobar por el Coordinador',
  THESIS1_APPROVED = 'Tesis 1: aprobada',
  THESIS2_PENDING_ADVISOR = 'Tesis 2: por aprobar por el Asesor',
  THESIS2_PENDING_COORDINATOR = 'Tesis 2: por aprobar por el Coordinador',
  THESIS2_APPROVED = 'Tesis 2: aprobada',
}